% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_microclimf_output.R
\name{microout_to_raster}
\alias{microout_to_raster}
\title{microout_to_raster}
\usage{
microout_to_raster(microout, r, epsg, climvars)
}
\arguments{
\item{microout}{output of run_micro or run_micro_dy}

\item{r}{raster tile with NA buffer of 1 row/column around the edge}

\item{epsg}{epsg code for output}

\item{climvars}{vector of climate variables to extract from microout (c('Tz', 'tleaf', 'T0', 'soilm', 'relhum', 'windspeed', 'raddir', 'raddif', 'radlw'))}
}
\value{
named list of spatRasters
}
\description{
converts output of microclimf models into raster format
}
