% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimf_parallel.R
\name{run_microclimf_parallel_update}
\alias{run_microclimf_parallel_update}
\title{run_microclimf_parallel_update}
\usage{
run_microclimf_parallel_update(
  tile,
  pai,
  dem,
  chm,
  habitats,
  soil,
  pai_seasonal,
  paia_seasonal,
  folden_seasonal,
  weather,
  precip,
  ll,
  reqhgt,
  epsg,
  n_years
)
}
\arguments{
\item{tile}{list of raster tiles to iterate over}

\item{pai}{spatRaster of PAI for entire study area}

\item{dem}{spatRaster of DEM for entire study area}

\item{chm}{spatRaster of CHM for entire study area}

\item{habitats}{spatRaster of habitat types (habitat type key in microclimf package) for entire study area}

\item{soil}{spatRaster of soil types for entire study area (soil type key in microclimf)}

\item{pai_seasonal}{array of PAI values as returned by lai_seasonality()}

\item{paia_seasonal}{array of PAI values above reqhgt as returned by paia_for_microf()}

\item{folden_seasonal}{array of foliage density at reqhgt as returned by folden_for_microf()}

\item{weather}{dataframe of hourly weather as returned by mcera5::extract_clim}

\item{precip}{vector of daily precipitation as returned by mcera5::extract_precip}

\item{ll}{dataframe of lat and lon values (see latlongfromSpatRaster())}

\item{reqhgt}{height at which to model microclimate}

\item{epsg}{epsg code}

\item{n_years}{number of years represented in weather data}
}
\value{
summ_rasts: spatRaster: each layer represents one summary variable
}
\description{
makes single spatraster with multiple bands for different summaries
}
