% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimf_prep.R
\name{seasonalRast_to_hourlyArray}
\alias{seasonalRast_to_hourlyArray}
\title{seasonalRast_to_hourlyArray}
\usage{
seasonalRast_to_hourlyArray(seasonalRast, tme)
}
\arguments{
\item{seasonalRast}{raster with nlyrs = 12. Each layer represents one month}

\item{tme}{vector of POSIX objects for the desired time period}
}
\value{
array of hourly seasonal values in array format where each layer corresponds to one hour.
}
\description{
Takes a spatRaster with each layer representing monthly seasonality of a different variable, replicates monthly seasonality to produce hourly, and converts to array.
}
