% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_microclimf_output.R
\name{summarise_microclimf}
\alias{summarise_microclimf}
\title{summarise_microclimf}
\usage{
summarise_microclimf(clim_rast, climdata, fun, months = NA, hours = NA)
}
\arguments{
\item{clim_rast}{spatRaster of a climate variable. Layers must represent times used to calculate microclimates}

\item{climdata}{climate dataframe that was used as input in microclimf}

\item{fun}{string: summary function to apply to clim_array (see fun options in terra::app)}

\item{months}{Months over which to apply summary function (1-12). If NA, summary function applied over all months}

\item{hours}{hours in day over which to apply summary function (0-23). If NA, summary function applied over all hours}
}
\value{
summary raster
}
\description{
summarizes clim_rast by designated time period. The function first subsets the data by months and hours
if provided, and then summarizes by day. All daily summaries are then averaged.
For example fun = max, will provide average daily maximum temperature
}
